+++ src/clients.rs
@@ -17,24 +17,20 @@ use dusk_wallet_core::{
     UnprovenTransaction, POSEIDON_TREE_DEPTH,
 };
 use flume::Sender;
+use futures::StreamExt;
+use phoenix_core::transaction::StakeData;
 use phoenix_core::{Crossover, Fee, Note};
 use poseidon_merkle::Opening as PoseidonOpening;
 
 use std::path::Path;
 use std::sync::{Arc, Mutex};
 
-use rusk_schema::{
-    ExecuteProverRequest, FindExistingNullifiersRequest, GetAnchorRequest,
-    GetOpeningRequest, GetStakeRequest, PreverifyRequest, PropagateMessage,
-    StctProverRequest, Transaction as TransactionProto, WfctProverRequest,
-};
-
 use self::sync::{sync_db, SYNC_INTERVAL_SECONDS};
 
 use super::block::Block;
 use super::cache::Cache;
 
-use super::rusk::{RuskNetworkClient, RuskProverClient, RuskStateClient};
+use crate::rusk::{RuskHttpClient, RuskRequest};
 use crate::store::LocalStore;
 use crate::Error;
 
@@ -48,24 +44,23 @@ const STCT_INPUT_SIZE: usize = Fee::SIZE
 const WFCT_INPUT_SIZE: usize =
     JubJubAffine::SIZE + u64::SIZE + JubJubScalar::SIZE;
 
+const TRANSFER_CONTRACT: &str =
+    "0100000000000000000000000000000000000000000000000000000000000000";
+const STAKE_CONTRACT: &str =
+    "0200000000000000000000000000000000000000000000000000000000000000";
+
 /// Implementation of the ProverClient trait from wallet-core
 pub struct Prover {
-    client: Mutex<RuskProverClient>,
-    state: Mutex<RuskStateClient>,
-    network: Mutex<RuskNetworkClient>,
+    state: RuskHttpClient,
+    prover: RuskHttpClient,
     status: fn(status: &str),
 }
 
 impl Prover {
-    pub fn new(
-        client: RuskProverClient,
-        state: RuskStateClient,
-        network: RuskNetworkClient,
-    ) -> Self {
+    pub fn new(state: RuskHttpClient, prover: RuskHttpClient) -> Self {
         Prover {
-            client: Mutex::new(client),
-            state: Mutex::new(state),
-            network: Mutex::new(network),
+            state,
+            prover,
             status: |_| {},
         }
     }
@@ -85,36 +80,23 @@ impl ProverClient for Prover {
         &self,
         utx: &UnprovenTransaction,
     ) -> Result<Transaction, Self::Error> {
-        let utx_bytes = utx.to_var_bytes();
-        let msg = ExecuteProverRequest { utx: utx_bytes };
-        let req = tonic::Request::new(msg);
-
         self.status("Proving tx, please wait...");
-        let mut prover = self.client.lock().unwrap();
-        let proof_bytes = prover.prove_execute(req).wait()?.into_inner().proof;
+        let utx_bytes = utx.to_var_bytes();
+        let prove_req = RuskRequest::new("prove_execute", utx_bytes);
+        let proof_bytes = self.prover.call(2, "rusk", &prove_req).wait()?;
         self.status("Proof success!");
-
-        self.status("Attempt to preverify tx...");
         let proof = Proof::from_slice(&proof_bytes).map_err(Error::Bytes)?;
         let tx = utx.clone().prove(proof);
         let tx_bytes = tx.to_var_bytes();
-        let tx_proto = TransactionProto {
-            version: 1,
-            r#type: 1,
-            payload: tx_bytes.clone(),
-        };
-        let msg = PreverifyRequest { tx: Some(tx_proto) };
-        let req = tonic::Request::new(msg);
-        let mut state = self.state.lock().unwrap();
-        state.preverify(req).wait()?;
+
+        self.status("Attempt to preverify tx...");
+        let preverify_req = RuskRequest::new("preverify", tx_bytes.clone());
+        let _ = self.state.call(2, "rusk", &preverify_req).wait()?;
         self.status("Preverify success!");
 
         self.status("Propagating tx...");
-        let msg = PropagateMessage { message: tx_bytes };
-        let req = tonic::Request::new(msg);
-
-        let mut net = self.network.lock().unwrap();
-        net.propagate(req).wait()?;
+        let propagate_req = RuskRequest::new("propagate_tx", tx_bytes);
+        let _ = self.state.call(2, "Chain", &propagate_req).wait()?;
         self.status("Transaction propagated!");
 
         Ok(tx)
@@ -139,14 +121,11 @@ impl ProverClient for Prover {
         writer.write(&address.to_bytes())?;
         writer.write(&signature.to_bytes())?;
 
-        let msg = StctProverRequest {
-            circuit_inputs: buf.to_vec(),
-        };
-        let req = tonic::Request::new(msg);
-
         self.status("Requesting stct proof...");
-        let mut prover = self.client.lock().unwrap();
-        let res = prover.prove_stct(req).wait()?.into_inner().proof;
+
+        let prove_req = RuskRequest::new("prove_stct", buf.to_vec());
+        let res = self.prover.call(2, "rusk", &prove_req).wait()?;
+
         self.status("Stct proof success!");
 
         let mut proof_bytes = [0u8; Proof::SIZE];
@@ -169,14 +148,9 @@ impl ProverClient for Prover {
         writer.write(&value.to_bytes())?;
         writer.write(&blinder.to_bytes())?;
 
-        let msg = WfctProverRequest {
-            circuit_inputs: buf.to_vec(),
-        };
-        let req = tonic::Request::new(msg);
-
         self.status("Requesting wfct proof...");
-        let mut prover = self.client.lock().unwrap();
-        let res = prover.prove_wfct(req).wait()?.into_inner().proof;
+        let prove_req = RuskRequest::new("prove_wfct", buf.to_vec());
+        let res = self.prover.call(2, "rusk", &prove_req).wait()?;
         self.status("Wfct proof success!");
 
         let mut proof_bytes = [0u8; Proof::SIZE];
@@ -203,7 +177,7 @@ pub struct StateStore {
 }
 
 struct InnerState {
-    client: RuskStateClient,
+    client: RuskHttpClient,
     cache: Arc<Cache>,
 }
 
@@ -212,8 +186,8 @@ use tokio::time::{sleep, Duration};
 impl StateStore {
     /// Creates a new state instance. Should only be called once.
     pub(crate) fn new(
+        client: RuskHttpClient,
         data_dir: &Path,
-        client: RuskStateClient,
         store: LocalStore,
     ) -> Result<Self, Error> {
         let cache = Arc::new(Cache::new(data_dir, &store)?);
@@ -298,18 +272,16 @@ impl StateClient for StateStore {
 
     /// Fetch the current anchor of the state.
     fn fetch_anchor(&self) -> Result<BlsScalar, Self::Error> {
-        let mut state = self.inner.lock().unwrap();
-
-        let msg = GetAnchorRequest {};
-        let req = tonic::Request::new(msg);
+        let state = self.inner.lock().unwrap();
 
         self.status("Fetching anchor...");
-        let res = state.client.get_anchor(req).wait()?.into_inner().anchor;
-        self.status("Anchor received!");
 
-        let mut bytes = [0u8; BlsScalar::SIZE];
-        bytes.copy_from_slice(&res);
-        let anchor = BlsScalar::from_bytes(&bytes)?;
+        let anchor = state
+            .client
+            .contract_query::<(), 0>(TRANSFER_CONTRACT, "root", &())
+            .wait()?;
+        self.status("Anchor received!");
+        let anchor = rkyv::from_bytes(&anchor).map_err(|_| Error::Rkyv)?;
         Ok(anchor)
     }
 
@@ -319,29 +291,20 @@ impl StateClient for StateStore {
         &self,
         nullifiers: &[BlsScalar],
     ) -> Result<Vec<BlsScalar>, Self::Error> {
-        let mut state = self.inner.lock().unwrap();
-
-        let null_bytes: Vec<_> =
-            nullifiers.iter().map(|s| s.to_bytes().to_vec()).collect();
-
-        let msg = FindExistingNullifiersRequest {
-            nullifiers: null_bytes,
-        };
-        let req = tonic::Request::new(msg);
+        let state = self.inner.lock().unwrap();
 
         self.status("Fetching nullifiers...");
-        let res = state
+        let nullifiers = nullifiers.to_vec();
+        let data = state
             .client
-            .find_existing_nullifiers(req)
-            .wait()?
-            .into_inner()
-            .nullifiers;
-        self.status("Nullifiers received!");
+            .contract_query::<_, 1024>(
+                TRANSFER_CONTRACT,
+                "existing_nullifiers",
+                &nullifiers,
+            )
+            .wait()?;
 
-        let nullifiers = res
-            .iter()
-            .map(|n| BlsScalar::from_slice(n))
-            .collect::<Result<Vec<_>, _>>()?;
+        let nullifiers = rkyv::from_bytes(&data).map_err(|_| Error::Rkyv)?;
 
         Ok(nullifiers)
     }
@@ -351,45 +314,52 @@ impl StateClient for StateStore {
         &self,
         note: &Note,
     ) -> Result<PoseidonOpening<(), POSEIDON_TREE_DEPTH, 4>, Self::Error> {
-        let mut state = self.inner.lock().unwrap();
-
-        let msg = GetOpeningRequest {
-            note: note.to_bytes().to_vec(),
-        };
-        let req = tonic::Request::new(msg);
+        let state = self.inner.lock().unwrap();
 
         self.status("Fetching opening notes...");
-        let res = state.client.get_opening(req).wait()?.into_inner().branch;
+
+        let data = state
+            .client
+            .contract_query::<_, 1024>(TRANSFER_CONTRACT, "opening", note.pos())
+            .wait()?;
+
         self.status("Opening notes received!");
 
-        let branch = rkyv::from_bytes(&res).map_err(|_| Error::Rkyv)?;
+        let branch = rkyv::from_bytes(&data).map_err(|_| Error::Rkyv)?;
         Ok(branch)
     }
 
     /// Queries the node for the amount staked by a key.
     fn fetch_stake(&self, pk: &PublicKey) -> Result<StakeInfo, Self::Error> {
-        let mut state = self.inner.lock().unwrap();
-
-        let msg = GetStakeRequest {
-            pk: pk.to_bytes().to_vec(),
-        };
-        let req = tonic::Request::new(msg);
+        let state = self.inner.lock().unwrap();
 
         self.status("Fetching stake...");
-        let res = state.client.get_stake(req).wait()?.into_inner();
+
+        let data = state
+            .client
+            .contract_query::<_, 1024>(STAKE_CONTRACT, "get_stake", pk)
+            .wait()?;
+
+        let res: Option<StakeData> =
+            rkyv::from_bytes(&data).map_err(|_| Error::Rkyv)?;
         self.status("Stake received!");
 
-        let amount = res.amount.map(|a| (a.value, a.eligibility));
-
+        let stake = res.ok_or(Error::NotStaked).map(
+            |StakeData {
+                 amount,
+                 reward,
+                 counter,
+             }| StakeInfo {
+                amount,
+                reward,
+                counter,
+            },
+        )?;
         let staking_address = pk.to_bytes().to_vec();
         let staking_address = bs58::encode(staking_address).into_string();
         println!("Staking address: {}", staking_address);
 
-        Ok(StakeInfo {
-            amount,
-            reward: res.reward,
-            counter: res.counter,
-        })
+        Ok(stake)
     }
 }