+++ src/clients/sync.rs
@@ -1,21 +1,23 @@
+use std::mem::size_of;
 use std::sync::Arc;
 
-use dusk_bytes::DeserializableSlice;
 use dusk_wallet_core::Store;
 use flume::Sender;
 use futures::StreamExt;
-use phoenix_core::Note;
-use rusk_schema::GetNotesRequest;
+use phoenix_core::transaction::{ArchivedTreeLeaf, TreeLeaf};
 
 use crate::{
-    clients::Cache, rusk::RuskStateClient, store::LocalStore, Error,
-    MAX_ADDRESSES,
+    clients::Cache, rusk::RuskHttpClient, store::LocalStore, Error,
+    RuskRequest, MAX_ADDRESSES,
 };
 
+use super::TRANSFER_CONTRACT;
+
+const RKYV_TREE_LEAF_SIZE: usize = size_of::<ArchivedTreeLeaf>();
 pub const SYNC_INTERVAL_SECONDS: u64 = 3;
 
 pub(crate) async fn sync_db(
-    client: &mut RuskStateClient,
+    client: &mut RuskHttpClient,
     store: &LocalStore,
     cache: &Cache,
     status: fn(&str),
@@ -35,12 +37,21 @@ pub(crate) async fn sync_db(
     let mut last_height = cache.last_height()?;
 
     status("Fetching fresh notes...");
-    let msg = GetNotesRequest {
-        height: last_height,
-        vk: vec![], // empty vector means *all* notes will be streamed
-    };
-    let req = tonic::Request::new(msg);
-    let mut stream = client.get_notes(req).await?.into_inner();
+
+    let req = rkyv::to_bytes::<_, 8>(&last_height)
+        .map_err(|_| Error::Rkyv)?
+        .to_vec();
+
+    let mut stream = client
+        .call_raw(
+            1,
+            TRANSFER_CONTRACT,
+            &RuskRequest::new("leaves_from_height", req),
+            true,
+        )
+        .await?
+        .bytes_stream();
+
     status("Connection established...");
 
     status("Streaming notes...");
@@ -48,18 +59,28 @@ pub(crate) async fn sync_db(
     status(format!("From block: {}", last_height).as_str());
     let _ = sender.send("Fetching notes from network".to_string());
 
-    while let Some(item) = stream.next().await {
-        let rsp = item?;
-        let note = Note::from_slice(&rsp.note)?;
+    // This buffer is needed because `.bytes_stream();` introduce additional
+    // spliting of chunks according to it's own buffer
+    let mut buffer = vec![];
+
+    while let Some(chunk) = stream.next().await {
+        buffer.extend_from_slice(&chunk?);
+        if buffer.len() < RKYV_TREE_LEAF_SIZE {
+            continue;
+        }
+        let TreeLeaf { block_height, note } =
+            rkyv::from_bytes(&buffer).map_err(|_| Error::Rkyv)?;
         let txn = cache.txn();
 
-        last_height = std::cmp::max(last_height, rsp.height);
+        buffer.clear();
+
+        last_height = std::cmp::max(last_height, block_height);
 
         for (ssk, vk, psk) in addresses.iter() {
             if vk.owns(&note) {
                 cache.insert(
                     psk,
-                    rsp.height,
+                    block_height,
                     (note, note.gen_nullifier(ssk)),
                     txn,
                 )?;